requirements-dev.txt
-r requirements.txt

django-debug-toolbar==3.5.0
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0

